Fix external help for BIFs
