+++ src/META-INF/plugin.xml
@@ -139,7 +139,7 @@
         <typedHandler implementation="org.intellij.erlang.editor.ErlangTypedHandler" id="erlang"/>
 
         <colorSettingsPage implementation="org.intellij.erlang.editor.ErlangColorSettingsPage"/>
-        <lang.documentationProvider language="Erlang" implementationClass="org.intellij.erlang.ErlangDocumentationProvider"/>
+        <lang.documentationProvider language="Erlang" implementationClass="org.intellij.erlang.documentation.ErlangDocumentationProvider"/>
 
         <sdkType implementation="org.intellij.erlang.sdk.ErlangSdkType"/>
         <compiler implementation="org.intellij.erlang.compilation.ErlangCompiler"/>
